+++ crates/native/src/pc.rs
@@ -105,12 +105,6 @@ impl Webrtc {
                         log::error!(
                             "Failed to remove audio track from sender: {e}",
                         );
-                    } else {
-                        let is_sending = self
-                            .audio_tracks
-                            .iter()
-                            .any(|t| !t.senders.is_empty());
-                        self.ap.set_output_will_be_muted(!is_sending);
                     }
                 }
                 _ => unreachable!(),
@@ -210,17 +204,7 @@ impl Webrtc {
                     sender.replace_video_track(None)
                 }
                 sys::MediaType::MEDIA_TYPE_AUDIO => {
-                    let result = sender.replace_audio_track(None);
-
-                    if result.is_ok() {
-                        let is_sending = self
-                            .audio_tracks
-                            .iter()
-                            .any(|t| !t.senders.is_empty());
-                        self.ap.set_output_will_be_muted(!is_sending);
-                    }
-
-                    result
+                    sender.replace_audio_track(None)
                 }
                 _ => unreachable!(),
             }